Add files via upload
The primary objective of this project is to develop a neural network model capable of accurately classifying images from the Fashion MNIST dataset, which contains 70,000 grayscale images of various clothing items. The project aims to explore the effectiveness of deep learning techniques in image classification and gain insights into the performance of different neural network architectures.